Group Morpheme Hypothesis by Phonetic Component
The MorphologicalAnalysis.best_morpheme_hypotheses returns
a hash of morpheme hypotheses indexed by their phonetic
component.

I created a MorphemeHypothesisPhoneticKey that serves as
the key for this hash.  It is a wrapper for the phonetic
component that hashes based on the string representation.

Raw phonetic component arrays of the same phone sequences
were getting different hash values in analyze-morphemes.
Strangely, I couldn't reproduce this problem in the test
script.