Alternate Semantic Equivalence Set Search
I was exhaustively searching for semantic equivalence
sets by enumerating the powerset of a set of morpheme
hypotheses for a given allophone.  Since this is
an exponential algorithm, it quickly becomes intractable
for even small data sets.

Now I search for semantic equivalence classes simply by
taking the intersection of the meanings of all the
morpheme hypotheses.  If there is no intersection, these
hypotheses are not considered.  Optionally we'll still do
a powerset search for small sets beneath a cutoff size.

The default cutoff size is POWERSET_SEARCH_CUTOFF=8 and
may be specified in analyze-morphemes.