Analysis Deep Copy; Variable Morpheme Objective Function
The MorphemeAnalysis object now supports deep copy.  This
enables doing morpheme hypothesis selection with a coverage
objective function.  It also can be the basis of forking
analyses in a beam search.

Pairs of words in alignments are tracked by the
MorphemeAnalysis object by index now, so
each_symmetric_pair is no longer used.