Align Morphemes
SurfaceMorphemes now align with other SurfaceMorphemes
that have the same meaning template (i.e. have the same
set of meaning features, though they may have different
values.)  The match rate counts morphemes as a single slot
and only gives credits to morpheme slots where morphemes
are aligned.  This inclines the search to discover a
partial ordering of morpheme types.

Sort analyses in beam search stringification by phonetic
coverage.  This is a better metric than the harmonic mean
of phonetic and semantic coverage.  (Though it's still
not a viable measure of the quality of an analysis.)

Also rename a function and clean up documentation in
analysis.rb.