Lazy Alignment; Global Scoring
Restructure the way alignment is done in
MorphologicalAnalysis.  Only when an analysis is created
are all the word pairs aligned.  After that alignments are
only done for those words that were changed by the
insertion of a morpheme.  This changed the main loop in
next_iteration, and also required the creation of a
two-way (alignments-to-word-indexes and word-indexes-to-
alignments) lookup table in AlignmentTable.

There is a cached global score for an analysis which the
beam uses in ranking.  Currently this is the match rate
for the alignments of all words.  The intuition is that
this is a more global metric, but it doesn't seem to work
as well as the match rate just for alignments involved in
new morpheme hypotheses.  I need a balance between these
two.

The Alignment object no longer subclasses the EditAlign
alignment but instead creates an EditAlign alignment at
constructor time, copies the relevant data out, then
discards it.  This allows us to deep_copy
MorphologicalAnalysis with a single Marshal load/dump
call.

Add length and [] to BeamSearch.